pyproject.toml
[tool.poetry]
name = "test_technique_itm"
version = "0.1.0"
description = "Ingest the data and insert it to Google Big Query"
authors = ["romainrz"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
pandas = ">=2.3.1,<3.0.0"
prefect = ">=3.4.9,<4.0.0"
azure-storage-blob = ">=12.26.0,<13.0.0"
google-cloud-bigquery = ">=3.35.0,<4.0.0"
pyarrow = ">=21.0.0,<22.0.0"
pandas-gbq = ">=0.29.2,<0.30.0"
google-cloud-bigquery-storage = ">=2.32.0,<3.0.0"

[tool.poetry.scripts]
runapp = "test_technique_itm.__main__:main"

[tool.pytest.ini_options]
pythonpath =["src"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
